+++ bittensor/__init__.py
@@ -36,6 +36,7 @@ __blocktime__ = 6
 import bittensor.axon
 import bittensor.config 
 import bittensor.neuron
+import bittensor.miner
 import bittensor.executor
 import bittensor.cli
 import bittensor.dendrite
