--- miners/IMAGE/ffnn_grunt/ffnn_grunt_config.yaml
@@ -1,38 +1,0 @@
-axon:
-  local_port: 8091
-  local_ip: 127.0.0.1
-  use_upnpc: False
-  max_workers: 10
-  max_gradients: 100
-receptor:
-  pass_gradients: true
-  timeout: 0.5
-  do_backoff: true
-  max_backoff: 100
-  key_dim: 100
-  stale_emit_filter: 10000
-  topk: 10
-metagraph:
-  network: akira # Test network
-  stale_emit_filter: 10000
-session:
-  learning_rate: 0.001
-  momentum: 0.98
-  batch_size_train: 8
-  batch_size_test: 64
-  log_interval: 10
-  sync_interval: 100
-  apply_remote_gradients: true
-  name: ffnn-grunt
-  accumulation_interval: 5
-  record_log: true
-nucleus:
-  max_workers: 5
-  queue_maxsize: 10000
-  queue_timeout: 5
-# Make sure the path of your coldkeyfile and hotkeyfile match!
-neuron:
-  coldkeyfile: ~/.bittensor/wallets/default/coldkeypub.txt
-  hotkeyfile: ~/.bittensor/wallets/default/hotkeys/default
-synapse:
-  target_dim: 10
