--- miners/IMAGE/mnist/mnist_config.yaml
@@ -1,27 +1,0 @@
-axon:
-  max_workers: 10
-  max_gradients: 100
-receptor:
-  pass_gradients: true
-  timeout: 0.5
-  do_backoff: true
-  max_backoff: 100
-  key_dim: 100
-  stale_emit_filter: 10000
-  topk: 10
-metagraph:
-  stale_emit_filter: 10000
-miner:
-  learning_rate: 0.001
-  momentum: 0.98
-  batch_size_train: 8
-  batch_size_test: 64
-  log_interval: 10
-  sync_interval: 10
-  record_log: false
-nucleus:
-  max_workers: 5
-  queue_maxsize: 10000
-  queue_timeout: 5
-synapse:
-  target_dim: 10
