--- miners/TEXT/gpt2_wiki/gpt2_wiki_config.yaml
@@ -1,53 +1,0 @@
-axon:
-  max_gradients: 100
-  max_workers: 10
-  use_upnpc: false
-metagraph:
-  stale_emit_filter: 10000
-miner:
-  accumulation_interval: 1
-  apply_remote_gradients: false
-  batch_size_train: 1
-  config_file: null
-  epoch_length: 500
-  learning_rate: 0.1
-  log_interval: 10
-  momentum: 0.98
-  n_epochs: 9223372036854775807
-  name: gpt2-wiki
-  record_log: false
-  root_dir: ~/.bittensor/miners/
-  sync_interval: 100
-neuron:
-  modality: 0
-nucleus:
-  max_workers: 5
-  queue_maxsize: 1000
-  queue_timeout: 5
-receptor:
-  do_backoff: true
-  max_backoff: 100
-  pass_gradients: true
-  timeout: 0.5
-router:
-  key_dim: 100
-  stale_emit_filter: 10000
-  topk: 10
-subtensor:
-  chain_endpoint: null 
-synapse:
-  activation_function: gelu_new
-  attn_pdrop: 0.1
-  embd_pdrop: 0.1
-  initializer_range: 0.02
-  layer_norm_epsilon: 1.0e-05
-  n_block_filter: 100
-  n_head: 1
-  n_inner: 8
-  n_layer: 2
-  resid_pdrop: 0.1
-  summary_activation: null
-  summary_first_dropout: 0.1
-  summary_proj_to_labels: true
-  summary_type: cls_index
-  summary_use_proj: true
