+++ miners/gpt2_genesis.py
@@ -15,6 +15,7 @@
 # THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION 
 # OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER 
 # DEALINGS IN THE SOFTWARE.
+
 """GPT2 Language Modelling miner
 
 This file demonstrates training the GPT2 neuron with language modelling.
@@ -48,49 +49,11 @@ from bittensor.utils.model_utils import ModelToolbox
 from synapses.gpt2 import GPT2Synapse
 from torch.nn.utils import clip_grad_norm_
 from transformers import AdamW
-from tqdm import tqdm
-from torch.utils.data.dataloader import DataLoader
-from datasets import load_dataset
+from qqdm import qqdm, format_str
+from bittensor.dataloaders.text_dataloader import GenesisTextDataloader
 
 
-
-class AdamCorpus():
-
-    def __init__(self, block_size: int, tokenizer=bittensor.__tokenizer__()):
-        self.block_size = block_size
-        self.tokenizer = tokenizer
-
-        self.lines = load_dataset('wikitext', 'wikitext-103-raw-v1')['train']
-
-    def __len__(self):
-        return len(self.lines) - self.block_size
-
-    def __getitem__(self, idx):
-        """ Returns a batch of sentences from text dataset.
-
-            Args:
-                idx: index of data input
-
-            Returns:
-                x
-        """
-
-        chunk = self.lines[idx:idx + self.block_size]['text']
-
-        dix = []
-        block_num=0
-        while block_num < self.block_size:
-            tokenized = self.tokenizer(chunk[block_num], padding=True, truncation=True)['input_ids']
-            for t in tokenized:
-                if block_num < self.block_size:
-                    dix.append(t)
-                    block_num += 1
-
-
-        x = torch.tensor(dix, dtype=torch.long)
-        return x
-
-class Miner():
+class Miner( bittensor.miner.Miner ):
 
     def __init__(self, config: Munch = None, **kwargs):
         if config == None:
@@ -98,9 +61,6 @@ class Miner():
         bittensor.config.Config.update_with_kwargs(config.miner, kwargs)
         Miner.check_config(config)
         self.config = config
-
-        # ---- Neuron ----
-        self.neuron = bittensor.neuron.Neuron(self.config)
 
         # ---- Model ----
         self.model = GPT2Synapse( self.config )
@@ -110,27 +70,21 @@ class Miner():
 
         # ---- Optimizer ----
         self.optimizer = self.configure_optimizers()
-
         self.lr = self.config.miner.learning_rate
+        self.training_loss = math.inf
+        self.best_train_loss = math.inf
+        self.rloss = math.inf
+        self.lloss = math.inf
+        self.dloss = math.inf
 
         # ---- Dataset ----
         # The Genesis Dataset:
         # The dataset used to train Adam and his first 100 children.
         # Here block size = sequence length.
-        self.dataset = AdamCorpus(self.model.get_block_size())
+        self.dataset = GenesisTextDataloader(self.config.miner.batch_size_train, self.model.get_block_size())
         self.tokens = 0
-
-        # ---- Logging ----
-        self.tensorboard = SummaryWriter(log_dir = self.config.miner.full_path)
-        if self.config.miner.record_log == True:
-            filepath = self.config.miner.full_path + "/{}_{}.log".format(self.config.miner.name, self.config.miner.trial_uid),
-            logger.add (
-                filepath,
-                format="{time:YYYY-MM-DD at HH:mm:ss} | {level} | {message}",
-                rotation="250 MB",
-                retention="10 days"
-            )
-                
+        super( Miner, self ).__init__( self.config, **kwargs )
+               
     @staticmethod
     def default_config() -> Munch:
         parser = argparse.ArgumentParser()
@@ -140,42 +94,68 @@ class Miner():
 
     @staticmethod
     def add_args(parser: argparse.ArgumentParser):
-        parser.add_argument('--miner.learning_rate', default=3e-2, type=float, help='Training initial learning rate.')
-        parser.add_argument('--miner.weight_decay', default=0.25, help='Model parameter weight decay.')
-        parser.add_argument('--miner.lr_decay', default=True, help='learning rate decay params: linear warmup followed by cosine decay to 10% of original.')
-        parser.add_argument('--miner.warmup_tokens', default=375e6, help='A linear LR warmup over the first miner.warmup_tokens tokens (default is 365 million)')
-        parser.add_argument('--miner.final_tokens', default=260e9, help='At what point we reach 10% of original LR')
-        parser.add_argument('--miner.num_workers', default=1, help='Number of workers for data loader.')
-
-        parser.add_argument('--miner.clip_gradients', default=1.0, type=float, help='Implement gradient clipping to avoid exploding loss on smaller architectures.')
-        parser.add_argument('--miner.n_epochs', default=int(sys.maxsize), type=int, help='Number of training epochs.')
-        parser.add_argument('--miner.epoch_length', default=500, type=int, help='Iterations of training per epoch')
-        parser.add_argument('--miner.batch_size_train', default=2, type=int, help='Training batch size.')
-        parser.add_argument('--miner.sync_interval', default=100, type=int, help='Batches before we sync with chain and emit new weights.')
-        parser.add_argument('--miner.log_interval', default=10, type=int, help='Batches before we log miner info.')
-        parser.add_argument('--miner.accumulation_interval', default=1, type=int, help='Batches before we apply acummulated gradients.')
-        parser.add_argument('--miner.apply_remote_gradients', default=True, type=bool, help='If true, neuron applies gradients which accumulate from remotes calls.')
-        parser.add_argument('--miner.root_dir', default='~/.bittensor/miners/', type=str,  help='Root path to load and save data associated with each miner')
-        parser.add_argument('--miner.name', default='gpt2-genesis', type=str, help='Trials for this miner go in miner.root / miner.name')
-        parser.add_argument('--miner.trial_uid', default=str(time.time()).split('.')[0], type=str, help='Saved models go in miner.root_dir / miner.name / miner.uid')
-        parser.add_argument('--miner.record_log', default=False, help='Record all logs when running this miner')
-        parser.add_argument('--miner.config_file', type=str, help='config file to run this neuron, if not using cmd line arguments.')
-        parser.add_argument('--debug', dest='debug', action='store_true', help='''Turn on bittensor debugging information''')
-        parser.set_defaults( debug=False )
+        parser.add_argument(
+            '--miner.learning_rate', 
+            default=3e-2, 
+            type=float, 
+            help='Training initial learning rate.'
+        )
+        parser.add_argument(
+            '--miner.weight_decay', 
+            default=0.25, 
+            type=float, 
+            help='Model parameter weight decay.'
+        )
+        parser.add_argument(
+            '--miner.lr_decay',
+            default=True,
+            type=bool,
+            help='learning rate decay params: linear warmup followed by cosine decay to 10%% of original.'
+        )
+        parser.add_argument(
+            '--miner.warmup_tokens',
+            default=375e6,
+            type=float,
+            help='A linear LR warmup over the first miner.warmup_tokens tokens (default is 365 million)'
+        )
+        parser.add_argument(
+            '--miner.final_tokens',
+            default=260e9,
+            type=float,
+            help='At what point we reach 10%% of original LR'
+        )
+        parser.add_argument(
+            '--miner.clip_gradients',
+            default=1.0,
+            type=float,
+            help='Implement gradient clipping to avoid exploding loss on smaller architectures.'
+        )
+        parser.add_argument(
+            '--miner.n_epochs', 
+            default=int(sys.maxsize), 
+            type=int, 
+            help='Number of training epochs.'
+        )
+        parser.add_argument(
+            '--miner.epoch_length', 
+            default=500, 
+            type=int, 
+            help='Iterations of training per epoch'
+        )
+        parser.add_argument(
+            '--miner.batch_size_train', 
+            default=2, 
+            type=int, 
+            help='Training batch size.'
+        )
+        parser.add_argument('--miner.name', default='gpt2_genesis', type=str, help='Trials for this miner go in miner.root / (wallet_cold - wallet_hot) / miner.name ')
         GPT2Synapse.add_args(parser)
-        bittensor.neuron.Neuron.add_args(parser)
+        bittensor.miner.Miner.add_args(parser)
 
     @staticmethod
     def check_config(config: Munch):
-        if config.debug:  bittensor.__log_level__ = 'TRACE'; logger.debug('DEBUG is ON')
-        else: logger.info('DEBUG is OFF') 
         assert config.miner.batch_size_train > 0, "batch_size_train must a positive value"
         assert config.miner.learning_rate > 0, "learning_rate must be a positive value."
-        full_path = '{}/{}/{}'.format(config.miner.root_dir, config.miner.name, config.miner.trial_uid)
-        config.miner.full_path = os.path.expanduser(full_path)
-        if not os.path.exists(config.miner.full_path):
-            os.makedirs(config.miner.full_path)
-
 
     def configure_optimizers(self):
         """
@@ -228,10 +208,10 @@ class Miner():
     def run (self):
 
         # ---- Subscribe ----
-        with self.neuron:
+        with self:
 
             # ---- Weights ----
-            self.row = self.neuron.metagraph.row.to(self.model.device)
+            self.row = self.metagraph.row.to(self.model.device)
 
             # --- Run state ---
             self.global_step = 0
@@ -240,7 +220,7 @@ class Miner():
             for self.epoch in range(self.config.miner.n_epochs):
 
                 # ---- Serve ----
-                self.neuron.axon.serve( self.model )
+                self.axon.serve( self.model )
 
                 # ---- Train Model ----
                 self.train()
@@ -252,37 +232,37 @@ class Miner():
                     continue
 
                 # ---- Emitting weights ----
-                self.neuron.metagraph.set_weights(self.row, wait_for_inclusion = True) # Sets my row-weights on the chain.
+                try:
+                    self.metagraph.set_weights(self.row, wait_for_inclusion = True) # Sets my row-weights on the chain.
 
-                # ---- Sync metagraph ----
-                self.neuron.metagraph.sync() # Pulls the latest metagraph state (with my update.)
-                self.row = self.neuron.metagraph.row.to(self.model.device)
-
-                # --- Epoch logs ----
-                #print(self.neuron.axon.__full_str__())
-                #print(self.neuron.dendrite.__full_str__())
-                #print(self.neuron.metagraph)
-
+                    # ---- Sync metagraph ----
+                    self.metagraph.sync() # Pulls the latest metagraph state (with my update.)
+                except:
+                    logger.error("Failed to set weights and sync metagraph! Could be  a connection  ")
+                
+                self.row = self.metagraph.row.to(self.model.device)
                 # ---- Update Tensorboard ----
-                self.neuron.dendrite.__to_tensorboard__(self.tensorboard, self.global_step)
-                self.neuron.metagraph.__to_tensorboard__(self.tensorboard, self.global_step)
-                self.neuron.axon.__to_tensorboard__(self.tensorboard, self.global_step)
+                self.dendrite.__to_tensorboard__(self.tensorboard, self.global_step)
+                self.metagraph.__to_tensorboard__(self.tensorboard, self.global_step)
+                self.axon.__to_tensorboard__(self.tensorboard, self.global_step)
 
                 # ---- Save best loss and model ----
-                if self.training_loss and self.training_loss < self.best_train_loss: #self.epoch % 10 == 0:
-                    if self.training_loss < self.best_train_loss:
+                if self.training_loss < self.best_train_loss: #self.epoch % 10 == 0:
                         self.best_train_loss = self.training_loss  # update best train loss
                         self.model_toolbox.save_model(
                             self.config.miner.full_path,
                             {
                                 'epoch': self.epoch,
                                 'model_state_dict': self.model.state_dict(),
-                                'loss': self.best_train_loss,
+                                'loss': self.best_train_loss/3,
                                 'optimizer_state_dict': self.optimizer.state_dict(),
+                                'rloss' : self.rloss,
+                                'lloss': self.lloss,
+                                'dloss': self.dloss,
                             }
                         )
                         self.tensorboard.add_scalar('Neuron/Train_loss', self.training_loss, self.global_step)
-                logger.info("This epoch's training loss: {}...Current best training loss: {}".format(self.training_loss, self.best_train_loss))
+                logger.info("This epoch's training losses: L-Loss: {:.2f} | R-Loss: {:.2f} | D-Loss: {:.2f} | avg: {:.2f} ... Current best average training loss: {:.2f}".format(self.lloss, self.rloss, self.dloss, self.training_loss/3, self.best_train_loss/3))
 
 
     def decay_learning_rate(self, batch):
@@ -312,86 +292,68 @@ class Miner():
         else:
             self.lr = self.config.miner.learning_rate
 
-
-    def reset_learning_rate(self, lr):
-        for param_group in self.optimizer.param_groups:
-            param_group['lr'] = lr
-
-    def shuffle_dataset_epoch_length(self):
-        """Shuffles the miner's dataset so we get a shuffled, randomized dataset
-        of length miner.epoch_length
-
-        Returns:
-            [list] : shuffled dataset of length miner.epoch_length
-        """
-
-        shuffled_dataset = []
-        loader = DataLoader(self.dataset, shuffle=True,
-                        batch_size=self.config.miner.batch_size_train,
-                        num_workers=self.config.miner.num_workers)
-
-
-        for it, batch in enumerate(loader):
-            shuffled_dataset.append(batch)
-            if it == self.config.miner.epoch_length:
-                break
-
-        return shuffled_dataset
-
     def get_lr(self):
         for param_group in self.optimizer.param_groups:
             return param_group['lr']
 
     # ---- Train Epoch ----
     def train(self):
-        self.training_loss = 0.0
 
         def run_epoch():
             self.model.train(True)
-            self.best_train_loss = math.inf
             losses = []
+            rlosses = []
+            llosses = []
+            dlosses = []
 
-            # Re-create dataloader every time we call train
-            # This way, since epoch_length < len(dataset), we can
-            # make sure that the dataset is randomly shuffled each time
             # we train for an epoch.
             logger.info("Preparing dataset batch...")
-            dataset = self.shuffle_dataset_epoch_length()
-            pbar = tqdm(enumerate(dataset), total=len(dataset))
-
-
-            #self.reset_learning_rate(self.config.miner.learning_rate)
+            # Set up the dataloader
+            dataloader = self.dataset.dataloader(self.config.miner.epoch_length)
+            pbar = qqdm(enumerate(dataloader), total=len(dataloader), desc=format_str('blue', f'Epoch Progress'))
             for it, (batch) in pbar:
+                # ---- Forward pass ----
                 batch = batch.to(self.model.device)
-                output = self.model.remote_forward(self.neuron, batch, training=True)
+                output = self.model.remote_forward(self, batch, training=True)
+
+                # ---- Backward pass ----
                 loss = output.local_target_loss + output.distillation_loss + output.remote_target_loss
                 loss.backward()
 
+                # ---- Gradient Step ----
                 clip_grad_norm_(self.model.parameters(), self.config.miner.clip_gradients)
                 self.optimizer.step()
                 self.optimizer.zero_grad()
-
                 self.decay_learning_rate(batch)
 
+                # Add losses up
                 losses.append(loss.item())
+                llosses.append(output.local_target_loss.item())
+                rlosses.append(output.remote_target_loss.item())
+                dlosses.append(output.distillation_loss.item())
 
-                 # ---- Train row weights ----
+                # ---- Train row weights ----
                 batch_weights = torch.mean(output.router.weights, axis = 0).to(self.model.device) # Average over batch.
                 self.row = (1 - 0.03) * self.row + 0.03 * batch_weights # Moving avg update.
                 self.row = F.normalize(self.row, p = 1, dim = 0) # Ensure normalization.
 
+                # ---- Logging ----
+                index = self.metagraph.state.index_for_uid[self.metagraph.uid]
                 pbar.set_infos({
                     'GS': colored('{}'.format(self.global_step), 'red'),
                     'LS': colored('{}'.format(it), 'blue'),
                     'Epoch': colored('{}'.format(self.epoch+1), 'green'),
-                    'Local loss': colored('{:.5f}'.format(output.local_target_loss.item()), 'red'),
-                    'Remote loss': colored('{:.5f}'.format(output.remote_target_loss.item()), 'blue'),
-                    'Distillation loss': colored('{:.5f}'.format(output.distillation_loss.item()), 'green'),
-                    'Learning Rate:': colored('{:e}'.format(self.lr), 'white'),
-                    'Axon': self.neuron.axon.__str__(),
-                    'Dendrite': self.neuron.dendrite.__str__(),
+                    'L-loss': colored('{:.5f}'.format(output.local_target_loss.item()), 'red'),
+                    'R-loss': colored('{:.5f}'.format(output.remote_target_loss.item()), 'blue'),
+                    'D-loss': colored('{:.5f}'.format(output.distillation_loss.item()), 'green'),
+                    'lr': colored('{:e}'.format(self.lr), 'white'),
+                    'nPeers': self.metagraph.n,
+                    'Stake(\u03C4)': float(self.metagraph.S[index]),
+                    'Rank(\u03C4)': float(self.metagraph.R[index]),
+                    'Incentive(\u03C4/block)': float(self.metagraph.I[index]),
+                    'Axon': self.axon.__str__(),
+                    'Dendrite': self.dendrite.__str__(),
                 })
-
                 self.tensorboard.add_scalar('Neuron/Rloss', output.remote_target_loss.item(), self.global_step)
                 self.tensorboard.add_scalar('Neuron/Lloss', output.local_target_loss.item(), self.global_step)
                 self.tensorboard.add_scalar('Neuron/Dloss', output.distillation_loss.item(), self.global_step)
@@ -399,11 +361,11 @@ class Miner():
 
 
             avg_loss = sum(losses) / len(losses)
-            self.training_loss = avg_loss
-            if avg_loss < self.best_train_loss:
-                self.best_train_loss = avg_loss
-                self.model_toolbox.save_model(self.config.miner.full_path,{'epoch': self.epoch, 'model_state_dict': self.model.state_dict(), 'loss': loss.item(), 'optimizer_state_dict': self.optimizer.state_dict()})
+            self.rloss = sum(rlosses) / len(rlosses)
+            self.lloss = sum(llosses) / len(llosses)
+            self.dloss = sum(dlosses) / len(dlosses)
 
+            self.training_loss = avg_loss
 
         run_epoch()
