+++ tests/miners/text/test_xlm.py
